refactor(adapter): adapter.web 에서 presentation으로 패키지 구조 변경
- 패키지 구조의 간소화와 명확성을 위해 inbound adapter인 web adapter 레이어를 presentation 레이어로 변경 (adapter.inbound.web.* -> adapter.web.* -> presentation.*)
- controller, dto, security, exception, common 패키지 이동
- 관련 import 경로 업데이트
