+++ backend/src/main/kotlin/org/setackle/backend/adapter/security/JwtAuthenticationEntryPoint.kt
@@ -3,7 +3,7 @@ package org.setackle.backend.adapter.security
 import com.fasterxml.jackson.databind.ObjectMapper
 import jakarta.servlet.http.HttpServletRequest
 import jakarta.servlet.http.HttpServletResponse
-import org.setackle.backend.adapter.web.common.ErrorResponse
+import org.setackle.backend.presentation.common.ErrorResponse
 import org.setackle.backend.common.exception.ErrorCode
 import org.slf4j.LoggerFactory
 import org.springframework.http.MediaType
