+++ backend/src/main/kotlin/org/setackle/backend/presentation/security/RequestValidationInterceptor.kt
@@ -1,4 +1,4 @@
-package org.setackle.backend.adapter.web.security
+package org.setackle.backend.presentation.security
 
 import jakarta.servlet.http.HttpServletRequest
 import jakarta.servlet.http.HttpServletResponse
