+++ backend/src/main/kotlin/org/setackle/backend/presentation/user/dto/LogoutAllDto.kt
@@ -1,4 +1,4 @@
-package org.setackle.backend.adapter.web.user.dto
+package org.setackle.backend.presentation.user.dto
 
 import org.setackle.backend.application.user.inbound.LogoutAllSessionsCommand
 import org.setackle.backend.application.user.inbound.LogoutAllSessionsResult
